+++ src/converter/index.ts
@@ -1,3 +1,2 @@
 export * from "./converterService";
 export * from "./ffmpegService";
-export * from "./fileLockService";
