+++ src/watcher/index.ts
@@ -1,1 +1,3 @@
+export * from "./fileLockService";
+export * from "./fileLockTypes";
 export * from "./fileWatcherService";
